Started work on v2 and refactoring drop structures
