+++ contract/src/stage1/function_call.rs
@@ -1,7 +1,6 @@
 use crate::*;
 
 /// Keep track of nft data 
-#[near_bindgen]
 #[derive(PanicOnDefault, BorshDeserialize, BorshSerialize, Serialize, Deserialize, Clone)]
 #[serde(crate = "near_sdk::serde")]
 pub struct FCData {
