+++ contract/src/stage1/mod.rs
@@ -1,2 +1,6 @@
 mod delete;
 mod drops;
+pub mod function_call;
+
+pub use drops::*;
+pub use function_call::*;
