+++ contract/src/stage2/mod.rs
@@ -1,8 +1,5 @@
 pub mod ft;
 pub mod nft;
-pub mod function_call;
 
 pub use ft::*;
 pub use nft::*;
-pub use function_call::*;
-
